Fix issue with multiple users building and running the code simultaneously

+++ q
@@ -2,5 +2,7 @@
 
 LOG_PATH="/home/hpe/NTUST-script/LOG"
 PID_FILE="$LOG_PATH/PID_FILE_NAME"
-
-sudo kill -9 `cat $PID_FILE`
+if [ -e $PID_FILE ]; then
+    sudo kill -9 `cat $PID_FILE`
+    rm $PID_FILE
+fi
